requirements.txt
python-telegram-bot[job-queue]>=20.0
pytz>=2024.1
flask>=2.0.0
requests>=2.31.0
schedule
